Consolidate Certificates, Composer, and DirectorPatch tabs into single-view one-click jailbreak
- Replace the three-tab UI with a unified Jailbreak view that runs all steps sequentially
- Auto-detect controller version and MAC via /api/v1/platform_status (replaces SSH/ARP)
- Auto-detect the Composer install path from Windows registry
- Auto-match and install management packs based on the detected controller version
  - Newer versions of Composer don't coincide with controller versions, and
    management packs must be installed separately.
- Make config patching and cert deployment idempotent (skip if unchanged)
- Add reboot confirmation prompt with /tmp marker for pending reboot detection
- Collect warnings and display them in a dedicated section at the end of the log output
