+++ UI/LoginDialog.cs
@@ -21,22 +21,33 @@ namespace Garry.Control4.Jailbreak.UI
         private void InitializeComponent()
         {
             Text = @"Control4 Customer Login";
-            Size = new System.Drawing.Size(350, 200);
+            Size = new System.Drawing.Size(350, 270);
             StartPosition = FormStartPosition.CenterParent;
             FormBorderStyle = FormBorderStyle.FixedDialog;
             MaximizeBox = false;
             MinimizeBox = false;
 
+            var explanationLabel = new Label
+            {
+                Text = @"SSH password authentication has been disabled by a system update. " +
+                       @"Your Control4 customer account credentials (from customer.control4.com) " +
+                       @"are needed to authenticate with the controller's local API in order to " +
+                       @"restore SSH access. Credentials are not stored.",
+                Location = new System.Drawing.Point(12, 12),
+                Size = new System.Drawing.Size(310, 70),
+                AutoSize = false
+            };
+
             var usernameLabel = new Label
             {
                 Text = @"Email:",
-                Location = new System.Drawing.Point(12, 15),
+                Location = new System.Drawing.Point(12, 85),
                 Size = new System.Drawing.Size(150, 23)
             };
 
             _usernameTextBox = new TextBox
             {
-                Location = new System.Drawing.Point(12, 35),
+                Location = new System.Drawing.Point(12, 105),
                 Size = new System.Drawing.Size(310, 23),
                 TabIndex = 0
             };
@@ -44,13 +55,13 @@ namespace Garry.Control4.Jailbreak.UI
             var passwordLabel = new Label
             {
                 Text = @"Password:",
-                Location = new System.Drawing.Point(12, 65),
+                Location = new System.Drawing.Point(12, 135),
                 Size = new System.Drawing.Size(150, 23)
             };
 
             _passwordTextBox = new TextBox
             {
-                Location = new System.Drawing.Point(12, 85),
+                Location = new System.Drawing.Point(12, 155),
                 Size = new System.Drawing.Size(310, 23),
                 UseSystemPasswordChar = true,
                 TabIndex = 1
@@ -59,7 +70,7 @@ namespace Garry.Control4.Jailbreak.UI
             _okButton = new Button
             {
                 Text = @"OK",
-                Location = new System.Drawing.Point(167, 125),
+                Location = new System.Drawing.Point(167, 195),
                 Size = new System.Drawing.Size(75, 23),
                 DialogResult = DialogResult.OK,
                 TabIndex = 2
@@ -68,7 +79,7 @@ namespace Garry.Control4.Jailbreak.UI
             _cancelButton = new Button
             {
                 Text = @"Cancel",
-                Location = new System.Drawing.Point(247, 125),
+                Location = new System.Drawing.Point(247, 195),
                 Size = new System.Drawing.Size(75, 23),
                 DialogResult = DialogResult.Cancel,
                 TabIndex = 3
@@ -76,6 +87,7 @@ namespace Garry.Control4.Jailbreak.UI
 
             _okButton.Click += OkButton_Click;
 
+            Controls.Add(explanationLabel);
             Controls.Add(usernameLabel);
             Controls.Add(_usernameTextBox);
             Controls.Add(passwordLabel);
